100 x2

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -6,12 +6,12 @@
  */
 int main(void)
 {
-	long int num;
-	long int i;
-	long int j;
-	long int p;
-	long int pnp;
-	long int comp = 612852475143;
+	long long int num;
+	long long int i;
+	long long int j;
+	long long int p;
+	long long int pnp;
+	long long int comp = 612852475143;
 	int b1 = 0;
 
 num = comp;
@@ -43,6 +43,6 @@ if (comp == 1)
 {
 num = i;
 } }
-printf("%ld\n", i - 1);
+printf("%lli\n", i - 1);
 return (0);
 }
